+++ NEWS.md
@@ -1,5 +1,13 @@
 # guideR (development version)
 
+**New features**
+
+* new experimental functions `tbl_maihda()`, `tbl_strata_info()` and 
+  `tbl_strata_predictions()()` to display the results of a MAIHDA analysis
+* new utility `add_glance_header()` to add a group header before glance
+  statistics added to a `gtsummary::tbl_regression()` table with 
+  `gtsummary::add_glance_table()`
+
 # guideR 0.10.0
 
 **New features**